refactor: make UsersController fully RESTful and align endpoint structure

+++ fridgeApi/src/main/java/com/example/fridgeapi/repositories/UsersRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<Users, BigInteger> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
     Optional<Users> findByEmail(String email);
 }
